requirements.txt
django==3.2.1
psycopg2-binary==2.8.6
djangorestframework==3.12.2
celery==4.4.6
django-debug-toolbar==3.2.4
drf-yasg==1.20.0
celery==4.4.6
redis==3.2.0